+++ packages/react-hooks/src/react-hooks.tsx
@@ -4,7 +4,7 @@ import {
   ShapeStream,
   ShapeStreamOptions,
 } from '@electric-sql/client'
-import React, { useCallback, useRef } from 'react'
+import React from 'react'
 import { useSyncExternalStoreWithSelector } from 'use-sync-external-store/with-selector.js'
 
 const streamCache = new Map<string, ShapeStream>()
@@ -55,7 +55,7 @@ export function getShape(shapeStream: ShapeStream): Shape {
   }
 }
 
-interface UseShapeResult {
+export interface UseShapeResult {
   /**
    * The array of rows that make up the Shape.
    * @type {{ [key: string]: Value }[]}
@@ -93,38 +93,40 @@ function parseShapeData(shape: Shape, shapeHash: string): UseShapeResult {
   }
 }
 
-const identity = (arg: unknown) => arg
+function identity<T>(arg: T): T {
+  return arg
+}
 
 interface UseShapeOptions<Selection> extends ShapeStreamOptions {
   selector?: (value: UseShapeResult) => Selection
 }
 
 export function useShape<Selection = UseShapeResult>({
-  selector = identity as never,
+  selector = identity as (arg: UseShapeResult) => Selection,
   ...options
 }: UseShapeOptions<Selection>): Selection {
   const shapeHash = sortedOptionsHash(options)
   const shapeStream = getShapeStream(options as ShapeStreamOptions)
   const shape = getShape(shapeStream)
 
-  const latestShapeData = useRef(parseShapeData(shape, shapeHash))
-  if (shapeHash !== latestShapeData.current.shapeHash) {
-    latestShapeData.current = parseShapeData(shape, shapeHash)
-  }
-  const getSnapshot = React.useCallback(() => latestShapeData.current, [])
-  const shapeData = useSyncExternalStoreWithSelector(
-    useCallback(
-      (onStoreChange) =>
-        shapeSubscribe(shape, () => {
-          latestShapeData.current = parseShapeData(shape, shapeHash)
-          onStoreChange()
-        }),
-      [shape, shapeHash]
-    ),
-    getSnapshot,
-    getSnapshot,
-    selector
-  )
+  const useShapeData = React.useMemo(() => {
+    let latestShapeData = parseShapeData(shape)
+    const getSnapshot = () => latestShapeData
+    const subscribe = (onStoreChange: () => void) =>
+      shapeSubscribe(shape, () => {
+        latestShapeData = parseShapeData(shape)
+        onStoreChange()
+      })
 
-  return shapeData
+    return () => {
+      return useSyncExternalStoreWithSelector(
+        subscribe,
+        getSnapshot,
+        getSnapshot,
+        selector
+      )
+    }
+  }, [shape, selector])
+
+  return useShapeData()
 }