+++ packages/sync-service/lib/electric/plug/delete_shape_plug.ex
@@ -49,7 +49,7 @@ defmodule Electric.Plug.DeleteShapePlug do
       # FIXME: This has a race condition where we accidentally create a snapshot & shape id, but clean
       #        it before snapshot is actually made.
       with {shape_id, _} <-
-             Shapes.get_or_create_shape_id(conn.assigns.shape_definition, conn.assigns.config),
+             Shapes.get_or_create_shape_id(conn.assigns.config, conn.assigns.shape_definition),
            :ok <- Shapes.clean_shape(shape_id, conn.assigns.config) do
         send_resp(conn, 202, "")
       end