+++ packages/sync-service/lib/electric/shape_cache/in_memory_storage.ex
@@ -1,6 +1,8 @@
 defmodule Electric.ShapeCache.InMemoryStorage do
   use Agent
 
+  require Electric.ShapeCache.LogChunker
+  alias Electric.ShapeCache.LogChunker
   alias Electric.ConcurrentStream
   alias Electric.Replication.LogOffset
   alias Electric.Shapes.Querying
@@ -14,12 +16,21 @@ defmodule Electric.ShapeCache.InMemoryStorage do
     snapshot_ets_table_name = Access.get(opts, :snapshot_ets_table, :snapshot_ets_table)
     log_ets_table_name = Access.get(opts, :log_ets_table, :log_ets_table)
 
+    chunk_checkpoint_ets_table_name =
+      Access.get(opts, :chunk_checkpoint_ets_table, :chunk_checkpoint_ets_table)
+
+    chunk_bytes_threshold =
+      Access.get(opts, :chunk_bytes_threshold, LogChunker.default_chunk_size_threshold())
+
     {:ok,
      %{
        snapshot_ets_table_base: snapshot_ets_table_name,
        log_ets_table_base: log_ets_table_name,
+       chunk_checkpoint_ets_table_base: chunk_checkpoint_ets_table_name,
        snapshot_ets_table: nil,
        log_ets_table: nil,
+       chunk_checkpoint_ets_table: nil,
+       chunk_bytes_threshold: chunk_bytes_threshold,
        shape_id: nil
      }}
   end
@@ -37,7 +48,13 @@ defmodule Electric.ShapeCache.InMemoryStorage do
           snapshot_ets_table:
             :ets.new(compiled_opts.snapshot_ets_table, [:public, :named_table, :ordered_set]),
           log_ets_table:
-            :ets.new(compiled_opts.log_ets_table, [:public, :named_table, :ordered_set])
+            :ets.new(compiled_opts.log_ets_table, [:public, :named_table, :ordered_set]),
+          chunk_checkpoint_ets_table:
+            :ets.new(compiled_opts.chunk_checkpoint_ets_table, [
+              :public,
+              :named_table,
+              :ordered_set
+            ])
         }
       end,
       name: name(compiled_opts.shape_id)
@@ -51,12 +68,14 @@ defmodule Electric.ShapeCache.InMemoryStorage do
   def for_shape(shape_id, compiled_opts) do
     snapshot_ets_table_name = Map.fetch!(compiled_opts, :snapshot_ets_table_base)
     log_ets_table_name = Map.fetch!(compiled_opts, :log_ets_table_base)
+    chunk_checkpoint_ets_table_name = Map.fetch!(compiled_opts, :chunk_checkpoint_ets_table_base)
 
     %{
       compiled_opts
       | shape_id: shape_id,
         snapshot_ets_table: :"#{snapshot_ets_table_name}-#{shape_id}",
-        log_ets_table: :"#{log_ets_table_name}-#{shape_id}"
+        log_ets_table: :"#{log_ets_table_name}-#{shape_id}",
+        chunk_checkpoint_ets_table: :"#{chunk_checkpoint_ets_table_name}-#{shape_id}"
     }
   end
 
@@ -105,14 +124,14 @@ defmodule Electric.ShapeCache.InMemoryStorage do
     {@snapshot_offset, stream}
   end
 
-  def get_log_stream(shape_id, offset, max_offset, opts) do
+  defp get_offset_indexed_stream(shape_id, offset, max_offset, offset_indexed_ets_table) do
     offset = storage_offset(offset)
     max_offset = storage_offset(max_offset)
 
-    :ets.tab2list(opts.log_ets_table)
+    :ets.tab2list(offset_indexed_ets_table)
 
     Stream.unfold(offset, fn offset ->
-      case :ets.next_lookup(opts.log_ets_table, {shape_id, offset}) do
+      case :ets.next_lookup(offset_indexed_ets_table, {shape_id, offset}) do
         :"$end_of_table" ->
           nil
 
@@ -126,6 +145,17 @@ defmodule Electric.ShapeCache.InMemoryStorage do
           {item, position}
       end
     end)
+  end
+
+  def get_log_stream(shape_id, offset, max_offset, opts) do
+    get_offset_indexed_stream(shape_id, offset, max_offset, opts.log_ets_table)
+  end
+
+  def get_chunk_end_log_offset(shape_id, offset, opts) do
+    get_offset_indexed_stream(shape_id, offset, offset, opts.chunk_checkpoint_ets_table)
+    |> Stream.map(fn {_, pos} -> LogOffset.new(pos) end)
+    |> Enum.take(1)
+    |> Enum.at(0)
   end
 
   def has_shape?(shape_id, opts) do
@@ -163,17 +193,44 @@ defmodule Electric.ShapeCache.InMemoryStorage do
     end)
   end
 
-  def append_to_log!(shape_id, log_items, opts) do
-    ets_table = opts.log_ets_table
+  def append_to_log!(shape_id, log_items, log_state, opts) do
+    log_ets_table = opts.log_ets_table
+    chunk_checkpoint_ets_table = opts.chunk_checkpoint_ets_table
+    chunk_bytes_threshold = opts.chunk_bytes_threshold
 
     log_items
-    |> Enum.map(fn log_item ->
-      offset = storage_offset(log_item.offset)
-      {{shape_id, offset}, Jason.encode!(log_item)}
-    end)
-    |> then(&:ets.insert(ets_table, &1))
+    |> Enum.flat_map_reduce(log_state, fn log_item, log_state ->
+      log_key = {shape_id, storage_offset(log_item.offset)}
+      json_log_item = Jason.encode!(log_item)
+      current_chunk_size = log_state.current_chunk_byte_size
 
-    :ok
+      case LogChunker.add_to_chunk(json_log_item, current_chunk_size, chunk_bytes_threshold) do
+        {:ok, new_chunk_size} ->
+          {
+            [{log_key, json_log_item}],
+            %{log_state | current_chunk_byte_size: new_chunk_size}
+          }
+
+        {:threshold_exceeded, new_chunk_size} ->
+          {
+            [
+              {log_key, json_log_item},
+              {{shape_id, storage_offset(log_item.offset)}, :checkpoint}
+            ],
+            %{log_state | current_chunk_byte_size: new_chunk_size}
+          }
+      end
+    end)
+    |> then(fn {items, log_state} ->
+      items
+      |> Enum.split_with(fn item -> match?({_, :checkpoint}, item) end)
+      |> then(fn {checkpoints, log_items} ->
+        :ets.insert(chunk_checkpoint_ets_table, checkpoints)
+        :ets.insert(log_ets_table, log_items)
+      end)
+
+      log_state
+    end)
   end
 
   def cleanup!(shape_id, opts) do