+++ packages/sync-service/lib/electric/shapes/snapshotter.ex
@@ -71,31 +71,31 @@ defmodule Electric.Shapes.Snapshotter do
 
   @doc false
   def query_in_readonly_txn(parent, shape_id, shape, db_pool, storage) do
-    Postgrex.transaction(db_pool, fn conn ->
-      OpenTelemetry.with_span("shape_cache.query_in_readonly_txn", [], fn ->
-        Postgrex.query!(
-          conn,
-          "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ READ ONLY",
-          []
-        )
+    Postgrex.transaction(
+      db_pool,
+      fn conn ->
+        OpenTelemetry.with_span("shape_cache.query_in_readonly_txn", [], fn ->
+          Postgrex.query!(conn, "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ READ ONLY", [])
 
-        %{rows: [[xmin]]} =
-          Postgrex.query!(conn, "SELECT pg_snapshot_xmin(pg_current_snapshot())", [])
+          %{rows: [[xmin]]} =
+            Postgrex.query!(conn, "SELECT pg_snapshot_xmin(pg_current_snapshot())", [])
 
-        GenServer.cast(parent, {:snapshot_xmin_known, shape_id, xmin})
+          GenServer.cast(parent, {:snapshot_xmin_known, shape_id, xmin})
 
-        # Enforce display settings *before* querying initial data to maintain consistent
-        # formatting between snapshot and live log entries.
-        Enum.each(Electric.Postgres.display_settings(), &Postgrex.query!(conn, &1, []))
+          # Enforce display settings *before* querying initial data to maintain consistent
+          # formatting between snapshot and live log entries.
+          Enum.each(Electric.Postgres.display_settings(), &Postgrex.query!(conn, &1, []))
 
-        stream = Querying.stream_initial_data(conn, shape)
+          stream = Querying.stream_initial_data(conn, shape)
 
-        GenServer.cast(parent, {:snapshot_started, shape_id})
+          GenServer.cast(parent, {:snapshot_started, shape_id})
 
-        # could pass the shape and then make_new_snapshot! can pass it to row_to_snapshot_item
-        # that way it has the relation, but it is still missing the pk_cols
-        Storage.make_new_snapshot!(shape_id, stream, storage)
-      end)
-    end)
+          # could pass the shape and then make_new_snapshot! can pass it to row_to_snapshot_item
+          # that way it has the relation, but it is still missing the pk_cols
+          Storage.make_new_snapshot!(shape_id, stream, storage)
+        end)
+      end,
+      timeout: :infinity
+    )
   end
 end